# scolaire-front/src/app/services/models/student-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StudentRequest {
  address: string;
  archive?: boolean;
  emailPerso?: string;
  emailPro: string;
  firstName: string;
  id?: number;
  lastName: string;
  phoneNumber: string;
}